+++ vscode-extension/src/allanguage/alFieldToolTipsLocation.ts
@@ -1,0 +1,4 @@
+export enum ALFieldToolTipsLocation {
+    page = 0,
+    table = 1
+}