+++ vscode-extension/src/codeactions/addFields/alAddQueryFieldsCodeCommand.ts
@@ -89,6 +89,6 @@ export class ALAddQueryFieldsCodeCommand extends ALBaseAddFieldsCodeCommand {
         }
         let source = writer.toString();
 
-        await this.insertSymbolContentAsync(symbol, source);
+        await this.insertSymbolContentAsync(symbol, source, range);
     }
 }